requirements.txt

# Core FastAPI and database dependencies
fastapi==0.115.12
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2

# Database drivers
psycopg2-binary==2.9.9  # For Supabase/PostgreSQL

# HTTP client and utilities
requests==2.31.0  # For API calls to image generation providers
httpx==0.25.0     # Alternative async HTTP client for Ollama
python-dotenv==1.1.0

# Image processing
Pillow==10.1.0    # For image processing and format conversion

# Async support
aiofiles==23.2.1  # For async file operations
asyncio           # Built into Python 3.7+

# Development and logging
python-json-logger==2.0.7
loguru==0.7.2

# Provider-specific dependencies
# All providers use standard HTTP requests, no additional deps needed